# backend/docs/swagger/complete-swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Vertex Markets API",
    "version": "1.0.0",
    "description": "Production-ready backend API for a modern Forex/Trading company website. Provides authentication, user management, CMS content management, markets, trading platforms, blog, contact management, newsletter subscriptions, careers/jobs, partner management, file uploads, admin dashboard, and global search capabilities.",
    "contact": {
      "name": "Vertex Markets Development Team"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://api.vertexmarkets.com",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from POST /api/v1/login"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "JWT access token stored in HTTP-only cookie"
      }
    },
    "schemas": {
...
      "DashboardStatistics": {
        "type": "object",
        "properties": {
          "users": {
            "type": "object",
            "properties": {
              "total": { "type": "integer", "example": 150 },
              "active": { "type": "integer", "example": 120 },
              "blocked": { "type": "integer", "example": 5 }
            }
          },
          "blog": {
            "type": "object",
            "properties": {
              "total": { "type": "integer", "example": 45 },
              "published": { "type": "integer", "example": 30 },
              "draft": { "type": "integer", "example": 15 }
            }
          },
          "cms": {
            "type": "object",
            "properties": {
              "totalSections": { "type": "integer", "example": 11 }
            }
          },
          "markets": {
            "type": "object",
            "properties": {
              "total": { "type": "integer", "example": 8 }
            }
          },
          "platforms": {
            "type": "object",
            "properties": {
              "total": { "type": "integer", "example": 5 }
            }
          },
          "contact": {
            "type": "object",
            "properties": {
              "total": { "type": "integer", "example": 200 },
              "new": { "type": "integer", "example": 25 },
              "replied": { "type": "integer", "example": 100 },
              "closed": { "type": "integer", "example": 75 }
            }
          },
          "newsletter": {
            "type": "object",
            "properties": {
              "total": { "type": "integer", "example": 500 },
              "active": { "type": "integer", "example": 450 },
              "unsubscribed": { "type": "integer", "example": 50 }
            }
          },
          "careers": {
            "type": "object",
            "properties": {
              "totalJobs": { "type": "integer", "example": 15 },
              "activeJobs": { "type": "integer", "example": 10 },
              "closedJobs": { "type": "integer", "example": 5 },
              "totalApplications": { "type": "integer", "example": 80 }
            }
          },
          "partners": {
            "type": "object",
            "properties": {
              "total": { "type": "integer", "example": 60 },
              "pending": { "type": "integer", "example": 15 },
              "approved": { "type": "integer", "example": 35 },
              "rejected": { "type": "integer", "example": 10 }
            }
          },
          "referrals": {
            "type": "object",
            "properties": {
              "total": { "type": "integer", "example": 120 }
            }
          },
          "uploads": {
            "type": "object",
            "properties": {
              "total": { "type": "integer", "example": 300 },
              "images": { "type": "integer", "example": 250 },
              "documents": { "type": "integer", "example": 50 }
            }
          }
        }
      },
      "SystemHealth": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "healthy" },
          "apiStatus": { "type": "string", "example": "running" },
          "serverStatus": { "type": "string", "example": "operational" },
          "uptime": { "type": "integer", "example": 86400 },
          "uptimeFormatted": { "type": "string", "example": "1d 0h 0m" },
          "totalModules": { "type": "integer", "example": 12 },
          "modules": { "type": "array", "items": { "type": "string" } },
          "registeredRoutes": { "type": "array", "items": { "type": "string" } },
          "totalRoutes": { "type": "integer", "example": 50 },
          "version": { "type": "string", "example": "1.0.0" },
          "environment": { "type": "string", "example": "development" },
          "nodeVersion": { "type": "string", "example": "20.0.0" },
          "platform": { "type": "string", "example": "win32" },
          "memoryUsage": {
            "type": "object",
            "properties": {
              "rss": { "type": "integer" },
              "heapTotal": { "type": "integer" },
              "heapUsed": { "type": "integer" },
              "external": { "type": "integer" }
            }
          },
          "startedAt": { "type": "string", "format": "date-time" }
        }
      },
      "DashboardResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "message": { "type": "string", "example": "Dashboard data fetched" },
          "data": {
            "type": "object",
            "properties": {
              "statistics": { "$ref": "#/components/schemas/DashboardStatistics" },
              "dashboardCards": {
                "type": "object",
                "properties": {
                  "growthSummary": { "type": "object" },
                  "pendingTasks": { "type": "object" },
                  "activeModules": { "type": "object" },
                  "systemHealth": { "type": "object" },
                  "totalRecords": { "type": "integer", "example": 1500 }
                }
              },
              "recentActivity": { "type": "array", "items": { "type": "object" } }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/dashboard": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Get full dashboard data",
        "description": "Returns aggregated statistics, dashboard cards, and recent activity. Admin only. Requires dashboard:read permission.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "query", "name": "limit", "schema": { "type": "integer", "default": 10 }, "description": "Limit for recent activity items" }
        ],
        "responses": {
          "200": {
            "description": "Dashboard data fetched",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DashboardResponse" }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/v1/dashboard/statistics": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Get aggregated statistics",
        "description": "Returns counts for users, blog, CMS, markets, platforms, contact, newsletter, careers, partners, and uploads. Admin only.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Statistics fetched",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "data": { "$ref": "#/components/schemas/DashboardStatistics" }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/v1/dashboard/recent": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Get recent items by type",
        "description": "Returns recent items filtered by type (users, blog, contacts, applications, partners, uploads) or all types if not specified. Admin only.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "query", "name": "type", "schema": { "type": "string", "enum": ["users", "blog", "contacts", "applications", "partners", "uploads"] }, "description": "Type of recent items to fetch" },
          { "in": "query", "name": "limit", "schema": { "type": "integer", "default": 10 }, "description": "Number of items to return" }
        ],
        "responses": {
          "200": { "description": "Recent items fetched" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/v1/dashboard/system": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Get system health information",
        "description": "Returns API status, server status, uptime, modules, routes, version, environment, and memory usage. Admin only.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "System health fetched",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "data": { "$ref": "#/components/schemas/SystemHealth" }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/v1/dashboard/activity": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Get aggregated activity feed",
        "description": "Returns a sorted feed of recent activities across users, blog, contacts, applications, partners, and uploads. Admin only.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "in": "query", "name": "limit", "schema": { "type": "integer", "default": 20 }, "description": "Number of activity items (max 100)" }
        ],
        "responses": {
          "200": { "description": "Activity feed fetched" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
...
  },
...
}